Fix comment handling when exiting diff view
Make sure that the revision row is updated with the current comment
count when after leaving the diff view.  Also, don't include draft
comments in the comment count (they are displayed separately).

When exiting the diff view with the esc key while inside of a comment
box, be sure to save that comment before leaving.

Change-Id: I16fe96dc37101d97317b19fdcc38be6729bed551