Ignore closed changes more.
Gertty previously synced any change it knew about and listed any change
not explicitly marked as reviewed in the unreviewed list. Update this to
ignore MERGED and ABANDONED changes when syncing and don't list these
changes in the unreviewed list either. Changes that are not open have
been reveiwed and closed and can be ignored by the user.

Note this is not symmetric currently with listing reviewed changes and
all changes. I think gertty should also learn to ignore and remove
MERGED changes from its DB entirely at some point. But not quite sure
how we want to do that. Might only be useful to do that after we add the
ability to view specific changes.

Change-Id: I951a687250b79eec7eca00f877cb1cb50e867c09