Removed closed changes from unreviewed list.
Previously any MERGED or ABANDONED changes showed up in the unreviewed
changes list for a project unless explicitly reviewed by the gertty
user. This is problematic because changes in these states are closed and
don't need further review. Remove them from the unreviewed list to cut
down on noise.

Note this is not symmetric currently with listing reviewed changes and
all changes. I think gertty should also learn to ignore and remove
MERGED changes from its DB entirely at some point. But not quite sure
how we want to do that. Might only be useful to do that after we add the
ability to view specific changes.

Change-Id: I951a687250b79eec7eca00f877cb1cb50e867c09