Remove MERGED changes from unreviewed list.
When listing the unreviewed changes that need review having the MERGED
changes show up is noisy. Remove them from the unreviewed change listing
as they have been merged and additional review is not necessary.

Note this is not symmetric currently with listing reviewed changes and
all changes. I think gertty should also learn to ignore and remove
MERGED changes from its DB entirely at some point. But not quite sure
how we want to do that. Might only be useful to do that after we add the
ability to view specific changes.

Change-Id: I951a687250b79eec7eca00f877cb1cb50e867c09