Sync parent changes
If a revision's parent commit is not in the db, searh for the
change associated with that parent and sync it.  Except, don't
do this if the revision in question is in a closed state (so
that we don't walk up the entire history).

This is in service of displaying change dependencies.

Change-Id: If1ecf7d82036255c742d4264f399b7f32ccd76b8