Alternate colors for light-background terminals
I use a white background on my terminal, and some of the default
color settings in the palette don't work with that. These changes
let me at least see everything, although things like cyan are still
a bit hard to read.

If you're interested in this as-is, that's great. If not, I can
rework it with a configuration option to let the user choose between
dark and light terminals in their config.

Change-Id: Ibee79e028bb15c20f8b27962a154491eb2d34e1d