Add hyperlinks
Add a 'link' commentlink substitution type that is a hyperlink
which will open the link in the user's web browser.  A built-in
configuration will apply this automaticlly to most http(s)?://
links (using a very simple and not entirely correct regex).

My current commentlink section in my config is as follows:

  commentlinks:
    - match: "^- (?P<job>.*?) (?P<url>.*?) : (?P<result>.*?) (?P<rest>.*)$"
      replacements:
        - link:
            text: "{job:<42}"
            url: "{url}"
        - text:
            color: "test-{result}"
            text: "{result} "
        - text: "{rest}"

In order to support the Zuul commentlink syntax.

Change-Id: Ifceee547c116fdcc15b50a2f73a0ddfe2e98af84