Handle multiple child revisions of same parent
Where there are multiple revisions of a child change that depend on
the same revision of the parent, we hit this:

    File ".../gertty/view/change.py", line 548, in refreshDependencies
      unseen_keys.remove(c.key)
  KeyError: 177

An example of this a change with this issue is that revision 1 of
I8cd0af9bd4d3669f436f169059e4b602d4d3036c is the parent of revisions 1 and 2
of If608302b6efacceb9ac3630b7ac402cf65ae6d68.

Handle this case by uniqifying the children list simply by using a mapping
of change key to subject.

Change-Id: I934c6df1bdf12bf87cc9e6827d38b92d04057f39