Support commits in commits
Gerrit represents changes to commits with a file change of /COMMIT_MSG
and magically offset file comments :/. In this patch I reproduce that
logic sufficiently well to match up on all the commits I have looked
at.

Since diffs vs the base should not show the old commit's commit
message, I have hinted to the diff function when to include the commit
message in the output.

Change-Id: I8adb9fa22b384cace88f114f770a3eb5d3a89f5c