Don't crash on comments on unchanged files
This can happen anytime a comment is made on an unchanged file,
either base -> pN, or pN -> pM, or even pN -> pN.

I've put a comment with what I think would be better behaviour, but
don't have time to research the right answer right now. Pushing
this up for folk that want non-crashing-gertty and/or pointers to
make the real fix approachable.

Change-Id: Ifb366dd1b77b1dafe2c5fb1f2ebc2035e6c4de4a