Reuse digest authentication state.
Digest authentication is stateful - a nonce is given by the server and
used for the next request, and the response includes the next nonce
and so forth. By creating a digest state each time we issue a GET /
POST, we were starting that process over leading to a new 401 -
effectively doubling our request counts.

Change-Id: Ib9fc674760f6aec66daf5244c8f0150596ea3ae3