Genericize the change list and add inter-change links
Make the change list display not only changes for a project, but
any arbitrary set of changes represented by a query.  This will
enable searching, inter-change linking, and custom dashboards.

The search syntax is extremly basic at the moment, but will be
extended to attempt to match the Gerrit webui search syntax as
closely as possible.

Use the new feature to implement a new kind of commentlink that
matches on Iabcdef style change-ids to "search" for matching
changes.

Change-Id: Icbcb166962fd96718ab56eb4118fc5ea43a4ba8b