Add a command to return to the change list
When viewing a change, map the key 'u' to return to the most recent
list of changes.  Normally this is the same as ESC, however, if
the user has followed a series of dependent changes, this will
pop the stack all the way back to the the list of changes to review.

Change-Id: I49f4f41f4b476e25f1bb83e92601597024c8c659