Fix crash on comments from undisplayed files
If there is a comment on a file that would not otherwise be displayed
because there is no diff, create an all-context diff for that file
so the comment may be displayed.

Also, update the number of comment lines to be displayed in the
context expansion button ("Expand X lines of context").

Change-Id: I6db509b42032b282dbbfdbee0d639f97b8b70da3