Add test results to top of change view
Add an option to commentlinks to indicate they are test results
and use that to collect results and display them at the top of the
change view.  The value of 'test-results' is used as the job key
when creating the table (so that only the most recent result for
a given job run on the latest patchset is included).  This is
particularly effective along with the hide-comment system to hide
the results from the comment stream.

Change-Id: I54a787914aab18ad2a2a5ca4cc79dd66add2c588