Reduce impact of check revisions task
Now that change refs are stored locally, the only time a local repo
should be missing revisions is when it is newly cloned.  Handle the
case where a user deletes the local repo automatically (when gertty
starts and a repo does not exist for a subscribed project, perform
a missing refs check on that project).

If the user knows something is wrong, they can supply the
--fetch-missing-refs flag at startup and an exhaustive check will
be run of all projects.

Otherwise, do nothing.

Change-Id: Iedebbe4b05ec1293b82c651587e42004444dc4a8