Don't modify status widgets outside of main thread
The sync module would indirectly modify the status widget (so that
the count of sync actions would be up to date).  If this happened
while urwid was drawing the screen, urwid may invalidate cached
information in mid-redraw, resulting in an exception and crash.
This updates the StatusHeader widget so that changes are merely
recorded to internal variables, and the actual widgets are updated
by a refresh method called from within the main loop.

Change-Id: I27043533d33d44aae31005670a1b727e9cc7bb90
Story: 260