Display version in help dialog title
It's nice to be able to tell what version you're running from within
the application itself, and the help dialog seems as good a place to
display it as any. Add it to the previously very short dialog title
line so as not to take up additional lines in the display.

Change-Id: Ie1ae713c4aeff5ae7136f1d5796f095f54db60aa