Use owner's username or email if display name is not set
If a change owner does not have a display name configured, the change
list doesn't display any name.

Rather than displaying nothing, attempt to use the username or email
instead if one of them is configured.

If no display name, username or email is set for the owner, default
to 'Anonymous Coward'.

Change-Id: Ifa2c56d89ed3d9b24e67643e460cacc133db90dd