Associate orphan messages with revision 1
Technically in the REST API, the revision number for a message is
optional.  In practice, all current Gerrit messages have a revision.
However, in reality (which is apparently different than practice)
Gerrit's Gerrit sometimes omits the revision number for some old
messages, even though the content of the message actually clearly
states "Patch Set 1".  To handle this case, just associate messages
with the first patchset so they will at least show up without error.

The main reason a message is associated with a patchset in the local
DB is for when we are writing a message.  So this solution is
probably good enough.

Change-Id: Ibf9d2a20277a89f24bb33d274342abce177f680e