Handle variable labels in change list
Different changes (even within a project) may have different review
labels associated with them.  When building the change listing screen,
iterate over all of the changes to get the superset of labels in
order to build the label columns on the right (the previous method
simply used the labels from the first change).

Change-Id: I77805eb3c0439cef86277c45b4261ad81aef3bbc