Do not limit the first sync to open review requests
While it's a default to only show open changes, it's possible to switch to
showing all of them, and that doesn't trigger a request for more aggressive
resync. This patch therefore fixes a regression introduced by I12049b4dcfa which
changed the behavior to never show changes which were already closed at the time
of the initial sync unless they are touched later on, no matter of what user's
setting is.

This might or might not be a desirable behavior -- the network traffic is surely
getting increased for the first, initial sync. However, the commit message of
the change in question does not speak about this at all, so I assume that it is
indeed a bug and was introduced by mistake.

Change-Id: Iaf1b7f27e697e4b125811d6abbd7d6e1a0eecbb9