Allow specifying the path to CA certificate bundle
The certificates are checked both by Gertty itself (for API calls) and
by the launched git processes. In theory, the server could be set up to
redirect to another HTTP server for Git calls (and in fact, the KDE's
Gerrit instance is set up to do just that). In that case, the CA bundle
file should contain PEM certificate chain of all the CAs for both Gerrit
and the webserver hosting the git repositories.

Change-Id: Id6af61c3710e4809c84b1edd054ab9b1959a60c3