Colorize votes on change list
The list of changes for a given project has two columns showing the max
vote for each change. Apply the min/max color on the votes, which with
the default pallette yields red for negative scores and green for
positive ones.

The palette has supports for slightly different color when the vote is
not the maximum allowed. Unfortunately the Gerrit API does not seem to
expose the label scores available to a project which make it impossible
to finely tune the color.  Assuming the color of the max/min scores will
work fine for most users.

Change-Id: I9f42be4cf41466a3500dcd98f7ad1cb6187b62a9