Release DB session thread lock earlier in syncs
For some of the sync tasks, the DB session context manager was holding
the Database object's thread lock for longer periods of time than was
necessary. This patch changes the sessionmaker in the Database object to
not expire objects on commit and not autoflush the session. This allows
the objects returned by a DB query to be used without running into an
UnboundExecutionError when trying to access attributes on the SQLAlchemy
models returned from a query after the session.commit() has been called.

We then modify some of the sync tasks to grab records and release the
thread lock on the DB session and then do their processing work instead
of holding the lock for the duration of the processing work.

Storyboard story: 2000086

Change-Id: Iae6d789409b88f4e3cb35ac9e4f839f8e4d79a97