Speed up the toggling of reviewed/hidden changes
When in the list changes view, if you typed 'v' or 'k' to mark a change
reviewed or hidden, a full refresh of the list view's contents,
including querying the database for all new records, would occur. For
the cases when we're in the Unreviewed changes list, toggling a change
to reviewed should simply remove that change from the listbox, to avoid
doing a full refresh. Same for the hidden toggle.

Storyboard story: 2000085

Change-Id: I76bb439cd5ce386da7d160d3160f36c769a2a99f