Always display full date
It's January 4. Just because a change was updated 5 days ago is no
reason to be merely telling me it was updated in '2014' - that makes it
impossible for me to tell the difference between a 5-day-old change and
a 369-day-old change.

(if I'd gotten around to fixing this on Jan 1 when it first bugged me, I
would have been talking about changes that were mere second old being
lumped in with changes almost a year old)

This change changes the date presentation for all dates in the Updated
column to be ISO-8601 date format, following the recomendations in that
most canonical of sources, http://xkcd.com/1179/

Change-Id: I4702ff5f226a3eb0cdeb258856686e6e0eaef41d