Always display full date
It's January 4. Just because a change was updated 5 days ago is no
reason to be merely telling me it was updated in '2014' - that makes it
impossible for me to tell the difference between a 5-day-old change and
a 369-day-old change.

(if I'd gotten around to fixing this on Jan 1 when it first bugged me, I
would have been talking about changes that were mere second old being
lumped in with changes almost a year old)

This change does several things, and I'm not sure if all of them are the
right thing to do.

* It removes the comparison of updated.year vs now.year; just because
  something was last year it doesn't follow that it was so long in the
  past that the exact date no longer matters.
* It always inserts %y ("Year without century as a decimal number") into
  the updated date field - but perhaps this is overkill?
* It changes the date presentation to day-month rather than month-day -
  now that the year is included, "22 Dec 14" makes more sense to me than
  "Dec 22 14". But maybe this is cultural bias - perhaps a better fix
  would be a configurable format string?

Change-Id: I4702ff5f226a3eb0cdeb258856686e6e0eaef41d