Always refresh the screen on pop
When going back a screen, always refresh the screen since something
may have changed while it was not displayed.  Also, refactor the
refresh methods to indicate whether something is of interest so
that if there is more than one refresh triggering event queued, we
still only refresh the screen once (if it is interested).

This should correct the problem where if you review a change and
go back to the change or project listings, the displayed numbers
are incorrect.

Change-Id: I97875831c4bd47f5e38992c38302b47809c785d5