Search: join tables when necessary
Some searches (eg, "owner:foo" or "project:bar") were performing
an implicit join on the primary key that joins them with the
change table in a 1:1 relationship.  If a user negated those,
then the result would be "not (a.key=b.key and a.value=foo)"
which is equal to "a.key!=b.key or a.value!=foo".  The partial
negation of the join criteria is definitely not what we want.

Instead, for the known 1:1 relationships that can be produced,
add the join criteria outside of the main expression iff those
tables are referenced.  Then the remaining part of the
condition (a.value == foo) can be optionally negated as normal.
Expressions for relationships other than 1:1 are handled by
subselects and so should not be impacted by this.

Also, I'm leaving a simple bit of debug code in the search
module for future use.

Change-Id: I92bb0a01f1768cb4f7cc9b3b28d54ebd309fa406