Fix approval sync
We're treating some fields in the db as if they have a unique
constraint, but there is no actual unique constraint in the db.
If, somehow, we ended up with multiple approvals for the same
user associated with a change, then the correct values may not
be displayed in the change approval table.

Correct this by detecting when we have multiple local entries
for an approval and remove entries subsequent to the first.
The first entry will remain and either be removed or have its
value updated as necessary.

Change-Id: Ifdc6273f1d5d4e83124ec30ff454c46e3e3af2d3