Set priority of initial change sync to normal
On startup, the initial update of all projects was set to high
priority.  That makes sense for most of the sync tasks at startup
(upload changes, etc).  But if you have left gertty to not sync
for any period of time, there could be quite a changes to sync.
With that happening at high priority, there is no way to find
a change and individually refresh it.  So move the initial sync
to low priority.  (Further syncs of all projects are at low
priority.)

Change-Id: I7cdffc39048ac4770bffefebe32d08371feb11be