Expand sample keymaps to ameliorate OSX features
OS X, for reasons of its own, swallows the CTRL+O keystroke, so OS X
users are likely to end up looking at this file to figure out how to fix
their keymap. This change adds an example of the mapping they're looking
for.

The secondary reason for documenting this is that none of the examples
show a multi-word name, so it took me a while to figure out the correct
name to use here. Adding an example makes it clearer how to use
multi-word names.

Change-Id: I7b95eea720864a5b18b9b5796bc30b4aebf57248