Fix syncing messages attached to draft revisions
If a revision is a draft, we will not be able to sync it.  Strangely,
the Gerrit API will still return review messages attached to that
draft (including "Foo uploaded patch set 1").  The gerrit webui doesn't
display those, so let's not bother storing/displaying them either.
The current code crashes when trying to save the message because the
revision doesn't exist.

Change-Id: I51617249c7240b199f0104c66b136922e36d6b7f
Story: 2000155