Hide webbrowser output
On my terminal, opening a web browser link always produces the output
"GLib-CRITICAL **: g_slice_set_config: assertion `sys_page_size == 0' failed"

This is apparently a bug in mozilla:

https://bugzilla.mozilla.org/show_bug.cgi?id=833117

However, the python webbrowser module could mask it by redirecting outputs.
In fact, it does so for some types of browsers, but apparently not xdg-open.
Create a new webbrowser subclass that redirects output and register it as
the appropriate way to use xdg-open so that the message is hidden.

Change-Id: I018673a2773a3bc21d7d7c1facccba91eafe1174