Only sync parent commit once
When a revision's parent commit is not known locally, it is added
to the sync queue.  However, if multiple revisions of a change have
the same parent (which is common) a sync task will be added for
each copy.  Instead, keep track of the parents of each revision
of a change and only sync each such commit once.

Change-Id: Ibe8fc80d87b338b6da9d529b1be38c7c4efb65f4