Right align line numbers
And change the column width to 5.

With left-aligned numbers and a column width of 4, the only time
numbers would abut the content was with >= 4 digit numbers.  However,
with right alignment we need to format 1 space less than we want to
leave room for in the general case.  If we format for 3 digit numbers
in a 4 column space, then we end up with a line number misalignment
between 999 and 1000.  That's reasonably common in our files.

Instead, format for 4 digit numbers in a 5 column space, which means
that numbers will be correctly aligned through 9999, and only on line
10000 will the column shift and the right side of the number abut the
content (note, the content itself will still not shift).  That seems
a reasonable compromise since no one will actually be able to load
a file that large using the web interface anyway.

Change-Id: If2b81a215dc6537ded5098ddb96f86d54be12633