Do not display InsecurePlatformWarning
The current requests module issues InsecurePlatformWarning on what
seems like every request.  Perform duplicate suppresion on all
warning messages.  Additionally, never display this message to the
user because there is almost nothing the user can do about it,
and annoying them is not going to help.  Log the message once.

Also move the InsecureRequestWarning suppression to use the same
new system.  Log it once if ssl verification is disabled.  Display
the warning if it is not disabled.

Change-Id: Icc84596dbc5f0e36a1af5c56215380b0f405613f