Add support for auth-type=form
I'm using a gerrit installation that uses neither 'basic' nor 'digest'
authentication methods, but instead redirects users to a custom login
form which sets a cookie to provide authentication.

This commit adds a new auth-type 'form', that automates logging in
through that form.

Currently, the path to the form is hardcoded to '/login' and the names
of the username and password fields to 'username' and 'password'
respectively. If preferred, I can add configuration options for each of
those too.

Change-Id: I3fff38b1be3f17c753787c9eb62ede0ce60a3ece