tox: Fix flake8 setup
This fixes the tox configuration for the flake8 target.

It excludes all existing flake8 issues, so the flake8 target succeeds.
In the following commits, I will fix these flake issues individually,
and remove the corresponding exclude from the list, so that every commit
passes the flake8 test.

It also adds the generated .tox folder to .gitignore.

Change-Id: I1f583d9e344b5af644f99ba2a94a09ecaff61176