Sync change when missing refs
The logic to decide from where to fetch refs is getting complicated
and depends on the full repo url and ref.  This information can be
different for each revision of a change, but is only needed in the
rare case that a local repo was damaged or removed.  Rather than
storing all of this information in the database for what should be
a rare event, simply sync any changes that have missing refs and
use the normal code path for deciding where to fetch refs.

Change-Id: I4308db51b4847163b255686a62d5bad4e6226b4d