Do not enqueue duplicate tasks
If a change requires a parent change to be synced, it might enqueue
a sync change task for each patchset with the same parent change.
This might result in many times the number of change syncs as
should be required.  Instead, if a task is already enqueued in
a given queue (for the same priority), do not enqueue that task.

The 'has' method on the queue is no longer necessary, so is removed.

Change-Id: I6ebd67b78f99142a401766bed645ae6968c63bf8