Add files table
In order to support searching for changes by file modified, keep
a local table of files modified in each revision.  In order to
populate the table without needing to re-sync every change, run
a diffstat on every revision during the migration.

Once established, the table will be populated during the normal
change sync.

Add support for the "file:" operator in searching.

Change-Id: Idbd44944405bdbc93af8bc3c513498e4425ba75b