Tell flake8 to ignore bin/ and lib/
This is not a pep8 fix, and doesn't enable pep8 style checks. It *does*
tell flake8 not to inspect files in bin/ or lib/

(But why do those directories even exist, I hear you say? Because I'm
running gertty from a venv at the root of the checkout. An equally valid
way to fix this would be to tell me to not do that)

Change-Id: I099fcbab8333df98f2986b6e604e8f1bbc8d827f