Add ability to review multiple changes at once
Add a process mark to the change list so that multiple changes may
be selected at once for further operations.  Allow multiple changes
to be reviwed at once by selecting them with the process mark and
then pressing the usual review command key.  The categories are
simply taken from the first change in the list for simplicity.
This should be fine most of the time, but if the changes have
different categories available to them, or the user has different
access levels, this may not behave as intended.

Change-Id: I04a790d91b27b270cf1269c7bcb39c12d857ab32