Show all held changes in held-changes view
I commented on a review while offline yesterday. When I came online,
someone had provided a -2. Being a more-negative comment than mine, this
held my change, and added a warning in the top-right to tell me I had a
held change.

However, I couldn't *see* the held change in held-changes view, because
it only shows me open changes - and this particular change was abandoned
after the -2.

This change makes held-changes view show all held changes,not just open
held changes.

An alternative here might be to not warn the user about held changes if
they have been abandoned; but that could be confusing if my comment had
been "I think this shouldn't be abandoned because"....

Change-Id: I5165f50af30b51ae29f2639dd5fd4185b02aa1fa