Supply a default query on search
When opening the search dialog from a change list, use the current
query as the default value.  This lets you iterate on a query
without needing to type or copy it repeatedly.  However, it does
mean that you need to clear out the field if you want a completely
different query.

Change-Id: I1f31db1cde0b86c5a7b214866f1df686dd901c7c