Fix syncing changes with comments on a missing file
If a change has comments on a file that is not in a patchset,
Gertty will fail to sync the change because local db entries for
files are only created if the file is actually in the revision.

If a comment appears for an unknown file, create an entry in
the db for that file.  Note, this will cause the change to appear
in "file:" queries that match that file.  It's not clear to me
that is incorrect.

Change-Id: I4916b85a5cbd72ced6682448b774752f3f6cc821