Fix multi-key handling at top level
The example for multi-key handling, setting quit to ":q", did not
work because the key buffer clearing at the project screen was
interfering with the key buffering for unhandled input at the top
level.  Correct this by only clearing the keybuffer in the project
screen if that screen handled a command.  Otherwise, let the
unhandled input handler clear it if necessary.

Change-Id: I104141ceaa87e6178ca74467a55cb7a944ad0128