Always sync a specfically queried change
If a change is opened by id and already exists locally, perform
a high-priority background sync of the change to make sure it's
up to date.  This should help avoid the problem where a user
returns to a change in an unsubscribed project and it appears out
of date.

If the change doesn't exist locally, it is synced in the foreground
so that case is already covered.

Change-Id: Idb5daa4309b712e39eb7cde56afcb94fdb41f179