Use diff long options and uncap GitPython
GitPython had a very incompatible change that breaks any call that
uses a single character option (e.g., -U) that requires no space
between the option and the value. So far, this breaks the diff API
where we used the -U option instead of the long option equivalent.
For history:

https://github.com/gitpython-developers/GitPython/issues/382

So far, I've only seen one place where this bites gertty. Since the
GitPython author has stated this will not be reverted or fixed, having
a permanent cap on that library seems harsh as we won't get any other
bug fixes.

This fixes the known/found areas where we are affected.

Change-Id: Iadb279234af2ea01fbff35dc629c01dae5a3195c