Do getChangeIDs() in batches
SQLite limits the number of items that can be passed in an IN query,
with the result that calling getChangeIDs() to determine which IDs
exist already in the DB will fail if a large number of IDs are passed.
To prevent this, split large groups of IDs into batches and perform
multiple queries that are each under the limit.

Change-Id: I3689097e3cf5f3c145a10ae057f4f545dd01f22e