Allow for a threaded query of the project list.
Some gerrit servers have so many projects that they fall over when you
try to query for the full list (as well as when passing normal
pagination options to try and limit the response).

This approach, while unorthodox, works.  It is disabled by default, but
when enabled, it queries for all projects based on the first letter of
their name.  To make it snappier, it utilizes a thread pool while
iterating over the list of possible prefixes.