Add an outdated flag for changes
If we detect that a change sync has failed, mark that change as
outdated in the local database so that if it has failed in such a
way that Gertty continues to process its queue and eventually stamp
a project or query as up-to-date, it can go back later and try to
sync the failed change again.

NB: This change contains a schema migration.

Change-Id: Ieb22a510a8096d77c12d30aefbedd36f7f3037b0