Set starred from CLI with --star option
Add a --star option which works like the --open option but instead
of opening the change in the running Gertty merely sets the starred
flag on it. This is useful if you use your starred changes for
ad-hoc priority reviewing and want to queue up additional reviews
out of band without impacting your present review context.

Change-Id: I160a3e4ecc9a270d27b4400e5c18394de6121dcd